requirements.txt
Django==5.2.4
djangorestframework==3.15.2
Pillow==10.4.0
sqlparse==0.5.3
python-decouple==3.8
openai
django-storages[boto3]>=1.14
boto3==1.28.60
